Kill warning signed, unsigned
git-svn-id: file:///home/jj/hercules.svn/trunk@5537 956126f8-22a0-4046-8f4a-272fa8102e63

+++ tapedev.c
@@ -979,7 +979,7 @@ int gettapetype_byname (DEVBLK *dev)
 
 #if defined(HAVE_REGEX_H) || defined(HAVE_PCRE)
 
-    for (i=0; i < arraysize( fmttab ); i++)
+    for (i=0; i < (int)arraysize( fmttab ); i++)
     {
         rc = regcomp (&regwrk, fmttab[i].fmtreg, REG_ICASE);
         if (rc < 0)
